backend/requirements.txt
asgiref==3.8.1
attrs==25.3.0
black==25.9.0
click==8.3.0
dj-database-url==3.0.1
Django==5.2.1
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
django-rest-knox==5.0.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
gunicorn==23.0.0
inflection==0.5.1
isort==7.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
Markdown==3.8
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-decouple==3.8
pytokens==0.2.0
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.26.0
setuptools==80.9.0
sqlparse==0.5.3
typing_extensions==4.14.1
uritemplate==4.2.0
wheel==0.45.1
